+++ src/components/button/button/index.js
@@ -1,0 +1,1 @@
+export * from './Button';
