+++ src/components/button/index.js
@@ -1,1 +1,7 @@
-export * from './Button';
+export * from "./button";
+
+export * from "./buttonnavigation";
+
+export * from "./buttontext";
+
+export * from "./pillbutton";
